backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.1.0
coreapi>=2.3.3  # обновить до последней версии
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-cors-headers==4.5.0
django-filter==23.3
django-templated-mail>=1.1.1
djangorestframework==3.16.0
djoser==2.1.0
drf-extra-fields>=3.4.0
gunicorn==20.1.0
idna==3.4
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
wheel==0.45.1
Pillow>=11.3.0  # обновить до последней версии
psycopg2-binary
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.7.0
sorl-thumbnail>=12.11.0  # обновить до последней версии
sqlparse==0.4.3
testfixtures==6.18.5
uritemplate==4.1.1
urllib3==1.26.15
